refactor: Prediction member init, return by value max_speed_.
The constructor of the Prediction pred BehavioPlanner class member variable
is called in a explicit way.
The  maxSpeed() function memeber of the same class returns the
maximum speed by value and not by reference anymore because
it's a private class memeber that has to be constant
for the whole program life.